move 16.handling-attachments over to samples/javascript_nodejs
