--- javascript_nodejs/16.handling-attachments/index.js
@@ -1,59 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-const { BotFrameworkAdapter } = require("botbuilder");
-const { BotConfiguration } = require("botframework-config");
-const path = require("path");
-const restify = require("restify");
-const { AttachmentsBot } = require("./bot");
-
-const CONFIG_ERROR = 1;
-
-// Bot name as defined in .bot file.
-// See https://aka.ms/about-bot-file to learn more about .bot files.
-const BOT_CONFIGURATION = 'attachments';
-
-// Read botFilePath and botFileSecret from .env file.
-// Note: Ensure you have a .env file and include botFilePath and botFileSecret.
-const ENV_FILE = path.join(__dirname, './.env');
-const env = require('dotenv').config({ path: ENV_FILE });
-
-// Create HTTP server.
-let server = restify.createServer();
-server.listen(process.env.port || process.env.PORT || 3978, function () {
-    console.log(`\n${server.name} listening to ${server.url}.`);
-    console.log(`\nGet Bot Framework Emulator: https://aka.ms/botframework-emulator.`);
-    console.log(`\nTo talk to your bot, open handling-attachments.bot file in the emulator.`);
-});
-
-// .bot file path.
-const BOT_FILE = path.join(__dirname, (process.env.botFilePath || ''));
-
-let botConfig;
-try {
-    // Read bot configuration from .bot file. 
-    botConfig = BotConfiguration.loadSync(BOT_FILE, process.env.botFileSecret);
-}
-catch (err) {
-    console.log(`Error reading bot file. Please ensure you have valid botFilePath and botFileSecret set for your environment.`);
-    process.exit(CONFIG_ERROR);
-}
-
-// Get bot endpoint configuration by service name.
-const endpointConfig = botConfig.findServiceByNameOrId(BOT_CONFIGURATION);
-
-// Create adapter. See https://aka.ms/about-bot-adapter to learn more adapters.
-const adapter = new BotFrameworkAdapter({
-    appId: endpointConfig.appId || process.env.MicrosoftAppID,
-    appPassword: endpointConfig.appPassword || process.env.MicrosoftAppPassword
-});
-
-// Create the AttachmentsBot.
-const attachmentsBot = new AttachmentsBot();
-
-// Listen for incoming requests.
-server.post('/api/messages', async (req, res) => {
-    await adapter.processActivity(req, res, async (context) => {
-        await attachmentsBot.onTurn(context);
-    });
-});
