+++ samples/javascript_nodejs/16.handling-attachments/.env
@@ -1,0 +1,2 @@
+botFilePath=attachments.bot
+botFileSecret=
