+++ samples/javascript_nodejs/16.handling-attachments/README.MD
