+++ samples/javascript_nodejs/16.handling-attachments/bot.js
