+++ samples/javascript_nodejs/16.handling-attachments/index.js
@@ -1,0 +1,71 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+const { BotFrameworkAdapter } = require('botbuilder');
+const { BotConfiguration } = require('botframework-config');
+const path = require('path');
+const restify = require('restify');
+const { AttachmentsBot } = require('./bot');
+
+// Read botFilePath and botFileSecret from .env file.
+// Note: Ensure you have a .env file and include botFilePath and botFileSecret.
+const ENV_FILE = path.join(__dirname, './.env');
+const env = require('dotenv').config({ path: ENV_FILE });
+
+// Get the .bot file path.
+// See https://aka.ms/about-bot-file to learn more about .bot files.
+const BOT_FILE = path.join(__dirname, (process.env.botFilePath || ''));
+let botConfig;
+try {
+    // Read bot configuration from .bot file.
+    botConfig = BotConfiguration.loadSync(BOT_FILE, process.env.botFileSecret);
+} catch (err) {
+    console.error(`\nError reading bot file. Please ensure you have valid botFilePath and botFileSecret set for your environment.`);
+    console.error(`\n - The botFileSecret is available under appsettings for your Azure Bot Service bot.`);
+    console.error(`\n - If you are running this bot locally, consider adding a .env file with botFilePath and botFileSecret.`);
+    console.error(`\n - See https://aka.ms/about-bot-file to learn more about .bot file its use and bot configuration.\n\n`);
+    process.exit();
+}
+
+// For local development configuration as defined in .bot file.
+const DEV_ENVIRONMENT = 'development';
+
+// Bot name as defined in .bot file or from runtime.
+const BOT_CONFIGURATION = (process.env.NODE_ENV || DEV_ENVIRONMENT);
+
+// Get bot endpoint configuration by service name.
+const endpointConfig = botConfig.findServiceByNameOrId(BOT_CONFIGURATION);
+
+// Create adapter. See https://aka.ms/about-bot-adapter to learn more adapters.
+const adapter = new BotFrameworkAdapter({
+    appId: endpointConfig.appId || process.env.MicrosoftAppID,
+    appPassword: endpointConfig.appPassword || process.env.MicrosoftAppPassword
+});
+
+// Create HTTP server.
+let server = restify.createServer();
+server.listen(process.env.port || process.env.PORT || 3978, function () {
+    console.log(`\n${server.name} listening to ${server.url}.`);
+    console.log(`\nGet Bot Framework Emulator: https://aka.ms/botframework-emulator.`);
+    console.log(`\nTo talk to your bot, open handling-attachments.bot file in the emulator.`);
+});
+
+// Catch-all for errors.
+adapter.onTurnError = async (turnContext, error) => {
+    // Write out errors to console log.
+    // NOTE: In production environment, you should consider logging this to Azure
+    //       application insights.
+    console.error(`\n [onTurnError]: ${ error }`);
+    // Send a message to the user.
+    await turnContext.sendActivity(`Oops. Something went wrong!`);
+};
+
+// Create the AttachmentsBot.
+const bot = new AttachmentsBot();
+
+// Listen for incoming requests.
+server.post('/api/messages', async (req, res) => {
+    await adapter.processActivity(req, res, async (turnContext) => {
+        await bot.onTurn(turnContext);
+    });
+});
