+++ samples/javascript_nodejs/16.handling-attachments/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "16.handling-attachments",
+  "name": "handling-attachments",
   "version": "1.0.0",
   "description": "Bot Builder v4 handling attachments sample",
   "main": "index.js",
@@ -13,12 +13,18 @@
   "license": "MIT",
   "dependencies": {
     "axios": "^0.18.0",
-    "botbuilder": "^4.0.0-preview1.1302",
-    "botframework-config": "^4.0.0-preview1.1302",
+    "botbuilder": "^4.0.5",
+    "botframework-config": "^4.0.5",
     "dotenv": "^6.0.0",
     "restify": "^7.2.1"
   },
   "devDependencies": {
+    "eslint": "^5.6.0",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-node": "^7.0.1",
+    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-standard": "^4.0.0",
     "nodemon": "^1.18.4"
   }
 }
