+++ samples/javascript_nodejs/16.handling-attachments/resources/architecture-resize.png
